requirements.txt
langchain-chroma
langchain-core
pymongo==4.12.1
python-dotenv
pdfplumber==0.11.7
tensorflow==2.19.0
transformers
fastembed
langchain_google_genai
lark
tf-keras
langchain
langchain-community
langchain-text-splitters
langchain-docling
langchain-groq
langchain-huggingface
langgraph
